+++ CHANGELOG.md
@@ -14,6 +14,7 @@
   - ex: in this example, "lf2m dps brd farm" wouldn't trigger an alert, but "lf2m dps hogger farm" would
 - New feature: Disable in instance
   - With this option selected, alerts will be toggled off while you are inside an instanced zone (determined via [IsInInstance](https://wow.gamepedia.com/API_IsInInstance))
+- `Escape` key will now close the alert frame
 
 ## 2.4.2